Moving registration methods to a different namespace; fixes #60; fixes #22
